### 2019-01-01 #38 @CameronDevine
Looking over the reviewer lists I have identified the following as good candidates.

From the JOSE list:
- @moorepants

From the JOSS list:
- @acolum
- @vsoch
