### 2019-01-16 #38 @labarba
:wave @moorepants — I assigned you your first paper to handle as JOSE editor! This is one is a good fit for you.
