### 2019-01-16 #38 @labarba
👋  @moorepants — After a couple of failed attempts, I've now assigned you your first JOSE paper to edit! This one is a good fit for you. Godspeed!
