### 2019-01-22 #38 @moorepants
I have the first reviewer. Chris Dembia is a PhD candidate at Stanford in the Neuromuscular Biomechanics Lab and a core developer on the Simbody and Opensim projects. He has also worked with Python and SymPy.
