### 2019-01-26 #38 @chrisdembia
Please address the following questions in the way you find most appropriate.

1. For what size problems do you expect the tools to function well? It might be good to document the problem size you expect. For example, it seems that the web interface might not properly display results for problems with hundreds of equations and variables.

2. What are some concrete intended use case? Will the software be used in a class at your university?

3. Are there other similar tools out there?

4. It seems your submission is missing some items listed in the [Reviewer Guidelines](http://jose.theoj.org/about#reviewer_guidelines):

a. "The online repository of the software or learning module needs to contain guidelines for potential contributors who may want to: submit changes, make improvements or report issues."
b. "Pedagogical soundness...The authors should briefly explain their design in the JOSE paper."
c. There should be clear guidelines for third-parties wishing to: Contribute to the software/module, Report issues or problems with the software/module, Seek support.

I have submitted a PR to your StateMint with additional comments and suggested changes.

Web interface
=============

The web interface has a beautiful design, and is, overall, easy to follow.

1. What does the SAVE button do? I entered a name but it wasn't obvious what happened after that. I eventually discovered that the "OPEN EXISTING" button showed the systems I had saved, but this was not obvious and I am still confused about where these files actually reside. What happens if I close the browser window? Do these saved files disappear?
2. The DOWNLOAD button provides an RND file. What is this file format? I haven't heard of it before and I didn't find any description of it.
3. If I enter incorrect input and hit "= CALCULATE", I get the flashing hexagons forever. If I scroll down, I see "We're sorry. The function has encountered an error". Ideally, this message would be shown immediately so that the user does not keep waiting unnecessarily.
4. It's unclear if the formatting of the "Results" pane will scale well to many equations.
5. Where do you document the "Upload a photo" functionality of the web interface? It's unclear what this does or what the photo should be.
6. For each example, consider a brief description of what system is being modeled.

WEB interface tutorial
The figures in the tutorial are well-designed. The choice of example is good. However, the second half of the tutorial is confusing and is not sufficiently clear for someone with a traditional system dynamics background and is not familiar with linear graph theory.
1. The tutorial would benefit from making clear which parts of the tutorial describe what a user must do to prepare input versus describing what the StateMint software does. It seems that the user must manually derive elemental, continuity, and compatibility equations, and the software uses these equations to provide state and output equations. This could be made clearer.
2. There are numbers on the figures in the Normal Tree section (e.g., 1, 2, 3, 4), but you do not describe the meaning of these numbers.
3. The writing would benefit from the use of the active voice instead of the passive voice.

Python
======
It's atypical to put all the code in __init__.py. Please create more descriptive Python files.

Example.ipynb
1. The first paragraph seems incorrect (there are 3 methods for using StateMint, not only 2).
2. Out[5] shows $(t)$ on its own. Placing $t$ in parentheses like this seems odd; is there an error?

Mathematica
===========
I installed a trial of Mathematica and ran the example without errors.

Please state the version of Mathematica with which you tested StateMint.
