### 2019-01-26 #38 @labarba
👋  @moorepants — Did you plan to have a second reviewer for this submission?

@chrisdembia — Thanks for providing these comments to the author… they can start working on revisions already, but with respect to the JOSE process, we're still in the PRE-REVIEW issue. When the editor has finished assigning the reviewer(s) they recruit for a submission, they start the review with a command here (`@whedon start review`), which opens a _new_ issue where the review is handled. That new issue will have a review checklist at the top. See examples by following the link to a past review from one of our published papers: http://jose.theoj.org
