### 2019-01-27 #38 @moorepants
Yes, I'm trying to get a second reviewer. It isn't clear to me if we can go to the actual review issue without that person yet. Should we?
