### 2019-01-27 #38 @labarba
Well ... if you start the review now, then adding a second reviewer on the review issue will be a manual process (edit the first post to copy and paste the checklist, manually add user to team).

You need to add the reviewer here, _then_ start the review to automatically get both the checklist for both reviewers and their permissions to edit.
