### 2019-01-29 #38 @labarba
JOSE published the [Pynamical](http://jose.theoj.org/papers/8e0d222df1d4e63dcf61bebcb643714a) package as one of its first papers. Maybe its author, @gboeing could be interested in contributing a review of the current submission?
