### 2019-01-30 #44 @whedon
[REVIEW]: StateMint: A Set of Tools for Determining Symbolic Dynamic System Models Using Linear Graph Methods
**Submitting author:** @CameronDevine (<a href="http://orcid.org/0000-0002-6579-111X">Cameron Devine</a>)
**Repository:** <a href="https://github.com/CameronDevine/StateMint" target ="_blank">https://github.com/CameronDevine/StateMint</a>
**Version:** 2.1.0
