### 2019-01-30 #44 @labarba
👋  @chrisdembia, @gboeing — Thank you for agreeing to review for JOSE! This is where the action happens: work your way through the review checklist, feel free to ask questions or post comments here, and also open issues in the submission repository as needed. Godspeed!
