### 2019-01-30 #44 @labarba
@CameronDevine — please bear in mind the comments that @chrisdembia posted on the Pre-Review issue: https://github.com/openjournals/jose-reviews/issues/38#issuecomment-457782845

@chrisdembia — feel free to re-post those comments here, if you wish.
