### 2019-01-30 #44 @moorepants
Yes, @chrisdembia please repost those comments here, so we keep things in one place.
