### 2019-02-14 CameronDevine/StateMint#9 @gboeing
document requirements
With regards to https://github.com/openjournals/jose-reviews/issues/44

Could you document the package's dependencies in a `requirements.txt` file?
