### 2019-02-14 CameronDevine/StateMint#10 @gboeing
software archive doi
With regards to https://github.com/openjournals/jose-reviews/issues/44

Is the software archived with a DOI (somewhere like zenodo) and documented accordingly in `paper.md`? I don't see it.
