### 2019-02-14 #44 @gboeing
Overall, the software looks useful and the web interface is very attractively designed. The paper is pretty clear and well-written. Regarding revisions/issues: rather than duplicating them all here, I'll start by saying that @chrisdembia's review covered most of the points I wanted to make. In addition, I have opened two issues in the software's repo https://github.com/CameronDevine/StateMint/issues/9 and https://github.com/CameronDevine/StateMint/issues/10.
