### 2019-02-14 #44 @moorepants
Fantastic! @chrisdembia and @gboeing, thank you both for you careful and thoughtful reviews.

@CameronDevine The reviewers' comments and requests seem appropriate. The ball is now in your court to address the reviews. Please do so and keep us up-to-date here on your progress or with any questions/comments you have.
