### 2019-02-19 a7c7690
added a try catch block in case the tracking.js file isn't loaded
