### 2019-02-20 72382b5
added a modal dialog for error messages
