### 2019-02-21 fb732ee
added a guide for contributors
