### 2019-02-21 742a784
removing a broken link from the contributing file
