### 2019-02-21 be3b8d2
added note about mathematica version to mathematica readme
