### 2019-02-21 18ce96b
added a description of each example system
