### 2019-02-21 b6970ef
fixed some test errors and added a test for the about button
