### 2019-02-21 2fbb125
moved the bulk of the python library to a new file.
