+++ python/StateMint/__init__.py
@@ -1,220 +1,1 @@
-'''
-This package provides a function to symbolically determine the differential equation describing the dynamics of a system. As inputs the `Solve` function takes the elemental and constraint equations of a system. These equations must be in the forms specified by Rowell and Wormley in "System Dynamics: An Introduction". A `tutorial <https://github.com/CameronDevine/StateMint/blob/master/tutorial.md>`_ is available which covers how to prepare equations for use with this code. There is also an `example <https://github.com/CameronDevine/StateMint/blob/master/python/Example.ipynb>`_ which covers the details of how to use this code. Convenience functions to convert the symbolic matrices to Numpy objects are also included.
-
-This code can be installed by running `pip install StateMint`, or one of the other interfaces can be used. Available are a Mathematica package, and a web `interface <http://statemint.stmartin.edu/>`_.
-'''
-
-import sympy
-
-t, s = sympy.symbols('t s')
-dummy = sympy.symbols('dummy')
-
-def sympify_vec(data):
-	return [sympy.sympify(el + '(t)') for el in data]
-
-def sympify_eqs(data):
-	if len(data) == 0:
-		return []
-	return [sympy.sympify('Eq(' + el.replace("'", "(t).diff(t)").replace('=', ',') + ')') for el in data]
-
-def make_matrix(eqs, vars, D = False):
-	M = sympy.zeros(len(eqs), len(vars))
-	for i in range(len(eqs)):
-		for j in range(len(vars)):
-			if D:
-				M[i,j] = sympy.simplify(sympy.diff(eqs[i].subs(vars[j].diff(t), dummy), dummy))
-			else:
-				M[i,j] = sympy.simplify(sympy.diff(eqs[i], vars[j]))
-	return M
-
-def make_vec(vars):
-	return [sympy.sympify(str(v).replace('(t)', '')) for v in vars]
-
-class output:
-	'''A container for the resulting differential equation.
-
-	This class is a container in which the output from `StateMint.Solve` is placed. This allows the results of the solution to be easily accessed as members of the class.
-	'''
-
-	A = None
-	'''sympy.Matrix: The state matrix, :math:`A`.
-
-	This matrix is found in state space formulations of dynamic systems such as, :math:`\dot{x}=Ax+Bu`.
-	'''
-
-	B = None
-	'''sympy.Matrix: The input matrix, :math:`B`.
-
-	This matrix is found in state space formulations of dynamics systems such as, :math:`\dot{x}=Ax+Bu`.
-	'''
-
-	C = None
-	'''sympy.Matrix: The output matrix, :math:`C`.
-
-	This matrix is found in a linear output equation of the form, :math:`y=Cx+Du`.
-	'''
-
-	D = None
-	'''sympy.Matrix: The feedthrough matrix, :math:`D`.
-
-	This matrix is found in a linear output equation of the form, :math:`y=Cx+Du`.
-	'''
-
-	E = None
-	'''sympy.Matrix: The time derivative input matrix, :math:`E`.
-
-	This matrix is found state space formulations of dynamic systems such as, :math:`\dot{x}=Ax+Bu+E\dot{u}`.
-	'''
-
-	F = None
-	'''sympy.Matrix: The time derivative feedthrough matrix, :math:`F`.
-
-	This matrix is found in linear output equations of the form, :math:`y=Cx+Du+F\dot{u}`.
-	'''
-
[…]
