### 2019-02-21 401ba8c
modified the travis ci file so that the web test isn't performed on pull requests
