+++ .travis.yml
@@ -21,6 +21,7 @@ script:
 jobs:
   include:
     - stage: web test
+      if: type != pull_request
       addons:
         chrome: stable
       python: 2.7
