### 2019-02-21 792373f
made the easy changes asked for by the reviewers.
