+++ web/HTML/tutorial/tutorial.md
@@ -10,7 +10,7 @@ This method of finding a differential equation of a dynamic system works with mu
 
 ### System Types
 
-In each energy domain a through-variable, $v$, and an across-variable, $f$, is defined.
+In each energy domain, a through-variable, $v$, and an across-variable, $f$, is defined.
 A through-variable is a variable that generally has the same value at each element terminal.
 It corresponds to a physical quantity that would be measured flowing through an element.
 An across-variable is a variable that generally has a different value at each element terminal.
@@ -27,12 +27,12 @@ A list of common system types and their through-variables and across-variables c
 
 ### Elemental Equations
 
-Within each energy domain there are multiple element types which can be used to model a system in that domain.
-Across energy domains these elements can be grouped into three distinct types.
+Within each energy domain, there are multiple element types which can be used to model a system in that domain.
+Across energy domains, these elements can be grouped into three distinct types.
 
 #### A Type Elements
 
-A type elements are energy storage elements which relates the rate of change of the across-variable to the through-variable of the element.
+A type elements are energy storage elements which relate the rate of change of the across-variable to the through-variable of the element.
 
 | System Type   | Element             | Elemental Equation         | Parameter
 | ------------- | ------------------- | -------------------------- | ---------
@@ -104,18 +104,18 @@ The following elements can be modeled as gyrators,
 Linear graph representations of systems use "nodes" and connecting "branches" to diagram a system.
 Each node represents an independent across-variable value.
 Branches are created for each element in the system.
-In non electrical systems A type elements always connect to a ground node.
+In non-electrical systems, A type elements always connect to a ground node.
 Arrows on branches point in the direction of decreasing across-variable.
 Arrows on transformers and gyrators always point towards ground.
 An example of a linear graph appears in the Example section of this tutorial.
 
 ### Normal Trees
 
-In order to find the primary and secondary variables a normal tree can be constructed.
+In order to find the primary and secondary variables, a normal tree can be constructed.
 This normal tree should consist of $N-1$ branches from the linear graph where $N$ is the number of nodes in the linear graph.
-If multiple ground nodes are present in the linear graph they should be counted as a single node.
-Since the normal tree must be a tree structure no loops may be created when constructing the normal tree.
-To construct the normal tree select branches in the following order.
+If multiple ground nodes are present in the linear graph, they should be counted as a single node.
+Since the normal tree must be a tree structure, no loops may be created when constructing the normal tree.
+To construct the normal tree, select branches in the following order.
 
 1. Across-variable sources
 2. A type elements
@@ -123,17 +123,20 @@ To construct the normal tree select branches in the following order.
 4. D type elements
 5. T type elements
 
-For transformers one branch must be selected, and for gyrators both or neither branches can be selected.
+For transformers, one branch must be selected.
+For gyrators, both or neither branches can be selected.
 The elements in the normal tree are termed normal tree branches, while the elements not in the normal tree are called normal tree links.
 
 ### Primary and Secondary Variables
 
-Once the normal tree has been created it is trivial to determine the primary and secondary variables. These are defined as,
+Once the normal tree has been created it is trivial to determine the primary and secondary variables.
+Primary variables are defined as,
 
 * Across-variables on normal tree branches and
 * Through-variables on normal tree links.
 
-For each element the secondary variable is the non primary variable. In other words the secondary variables are,
+For each element, the secondary variable is the non-primary variable.
+In other words the secondary variables are,
 
[…]
