### 2019-02-22 605b608
updated the tutorial as requested by the reviewers
