+++ tutorial.md
@@ -10,7 +10,7 @@ This method of finding a differential equation of a dynamic system works with mu
 
 ### System Types
 
-In each energy domain a through-variable, ![$v$](http://latex.codecogs.com/svg.latex?v), and an across-variable, ![$f$](http://latex.codecogs.com/svg.latex?f), is defined.
+In each energy domain, a through-variable, ![$v$](http://latex.codecogs.com/svg.latex?v), and an across-variable, ![$f$](http://latex.codecogs.com/svg.latex?f), is defined.
 A through-variable is a variable that generally has the same value at each element terminal.
 It corresponds to a physical quantity that would be measured flowing through an element.
 An across-variable is a variable that generally has a different value at each element terminal.
@@ -27,12 +27,12 @@ A list of common system types and their through-variables and across-variables c
 
 ### Elemental Equations
 
-Within each energy domain there are multiple element types which can be used to model a system in that domain.
-Across energy domains these elements can be grouped into three distinct types.
+Within each energy domain, there are multiple element types which can be used to model a system in that domain.
+Across energy domains, these elements can be grouped into three distinct types.
 
 #### A Type Elements
 
-A type elements are energy storage elements which relates the rate of change of the across-variable to the through-variable of the element.
+A type elements are energy storage elements which relate the rate of change of the across-variable to the through-variable of the element.
 
 | System Type   | Element             | Elemental Equation         | Parameter
 | ------------- | ------------------- | -------------------------- | ---------
@@ -99,23 +99,30 @@ The following elements can be modeled as gyrators,
 | Hydraulic Ram     | ![$-1/A$](http://latex.codecogs.com/svg.latex?-1/A) | Piston Area, ![$A$](http://latex.codecogs.com/svg.latex?A)
 | Displacement Pump | ![$-1/D$](http://latex.codecogs.com/svg.latex?-1/D) | Pump Displacement, ![$D$](http://latex.codecogs.com/svg.latex?D)
 
+#### Sources
+
+The final type of element are the source elements.
+Source elements come in two types, across-variable sources and through-variable sources.
+These sources allow power to be injected into the system at a given across or through-variable time dependent level.
+
 ### Linear Graphs
 
 Linear graph representations of systems use "nodes" and connecting "branches" to diagram a system.
-Each node represents an independent across-variable value.
-Branches are created for each element in the system.
-In non electrical systems A type elements always connect to a ground node.
+Each node represents an independent across-variable value and is drawn as a dot, or a small circle.
+Branches are created for each element in the system, where power flows between the nodes.
+Branches are drawn as a line between two nodes with an arrow in the direction of the decreasing across-variable.
+In non-electrical systems, A type elements always connect to a ground node.
 Arrows on branches point in the direction of decreasing across-variable.
 Arrows on transformers and gyrators always point towards ground.
 An example of a linear graph appears in the Example section of this tutorial.
 
 ### Normal Trees
 
-In order to find the primary and secondary variables a normal tree can be constructed.
+In order to find the primary and secondary variables, a normal tree can be constructed.
 This normal tree should consist of ![$N-1$](http://latex.codecogs.com/svg.latex?N-1) branches from the linear graph where ![$N$](http://latex.codecogs.com/svg.latex?N) is the number of nodes in the linear graph.
-If multiple ground nodes are present in the linear graph they should be counted as a single node.
-Since the normal tree must be a tree structure no loops may be created when constructing the normal tree.
-To construct the normal tree select branches in the following order.
[…]
