### 2019-02-22 44f66e4
updated the sources section of the tutorial
