### 2019-02-22 33eb7c6
updated the tutorial based on the reviewers suggestions
