### 2019-02-23 04f4a1c
finished modifications of the paper
