### 2019-02-23 #44 @CameronDevine
I have addressed the issues in @chrisdembia's pull request, but am waiting to close it until I get confirmation I have sufficiently answered his questions. I have a few more items to address in @chrisdembia's comment on this issue. After that I will work on @gboeing's issues.
