### 2019-02-24 2023f37
tripled the amount of time calculations can run in the web interface
