### 2019-02-24 #44 @CameronDevine
@chrisdembia Below are my responses to your comments.

> 1. For what size problems do you expect the tools to function well? It might be good to document the problem size you expect. For example, it seems that the web interface might not properly display results for problems with hundreds of equations and variables.

The question about the web interface is addressed below. Doing a couple tests I found that solving for the state equation of N independent RC filters is O(N). However, solving N cascaded RC filters was O(e^N^2). So far I have not encountered a problem that was not able to be solved in a reasonable amount of time, except those which I fabricated to test the performance.

> 2. What are some concrete intended use case? Will the software be used in a class at your university?

Prof. Rico Picone at St. Martin's University and Prof. Joseph Garbini at The University of Washington are both planning to use this tool in their system dynamics and control systems classes. This tool will allow students to more easily complete homework problems because they will not have to perform as much algebra.

> 3. Are there other similar tools out there?

I believe there are similar tools available for use with Bond Graphs, but I do not know of any for use with Linear Graphs.

> 4. It seems your submission is missing some items listed in the [Reviewer Guidelines](http://jose.theoj.org/about#reviewer_guidelines):
>
> a. "The online repository of the software or learning module needs to contain guidelines for potential contributors who may want to: submit changes, make improvements or report issues."

This has been added.

> b. "Pedagogical soundness...The authors should briefly explain their design in the JOSE paper."

Would the addition to the paper of an explanation of the algorithm used to find the differential equations be enough to satisfy this requirement?

> c. There should be clear guidelines for third-parties wishing to: Contribute to the software/module, Report issues or problems with the software/module, Seek support.

This has been added.

# Web interface

> 1. What does the SAVE button do? I entered a name but it wasn't obvious what happened after that. I eventually discovered that the "OPEN EXISTING" button showed the systems I had saved, but this was not obvious and I am still confused about where these files actually reside. What happens if I close the browser window? Do these saved files disappear?

These files are saved in your browser as cookies. They are set to expire after 10 years, but they may be deleted before that happens. I understand how this could be confusing so I added a note about this in the save dialog.

> 2. The DOWNLOAD button provides an RND file. What is this file format? I haven't heard of it before and I didn't find any description of it.

This file is actually a JPEG image with the system data added as metadata. The RND name was created back when we were calling this tool by its initial name, State Model RnD. I would welcome any suggestions on what file extension to use, and where to document this.

> 3. If I enter incorrect input and hit "= CALCULATE", I get the flashing hexagons forever. If I scroll down, I see "We're sorry. The function has encountered an error". Ideally, this message would be shown immediately so that the user does not keep waiting unnecessarily.

This has been fixed.

> 4. It's unclear if the formatting of the "Results" pane will scale well to many equations.

A system of 15 independent RC filters can be solved using this [link](https://statemint.stmartin.edu/?%7B%22InVars%22%3A%20%22Vs1%2C%20Vs2%2C%20Vs3%2C%20Vs4%2C%20Vs5%2C%20Vs6%2C%20Vs7%2C%20Vs8%2C%20Vs9%2C%20Vs10%2C%20Vs11%2C%20Vs12%2C%20Vs13%2C%20Vs14%2C%20Vs15%22%2C%20%22StVarElEqns%22%3A%20%22Vc1%27%3DIc1/C1%2C%5CnVc2%27%3DIc2/C2%2C%5CnVc3%27%3DIc3/C3%2C%5CnVc4%27%3DIc4/C4%2C%5CnVc5%27%3DIc5/C5%2C%5CnVc6%27%3DIc6/C6%2C%5CnVc7%27%3DIc7/C7%2C%5CnVc8%27%3DIc8/C8%2C%5CnVc9%27%3DIc9/C9%2C%5CnVc10%27%3DIc10/C10%2C%5CnVc11%27%3DIc11/C11%2C%5CnVc12%27%3DIc12/C12%2C%5CnVc13%27%3DIc13/C13%2C%5CnVc14%27%3DIc14/C14%2C%5CnVc15%27%3DIc15/C15%22%2C%20%22OtherElEqns%22%3A%20%22IR1%3DVR1/R1%2C%5CnIR2%3DVR2/R2%2C%5CnIR3%3DVR3/R3%2C%5CnIR4%3DVR4/R4%2C%5CnIR5%3DVR5/R5%2C%5CnIR6%3DVR6/R6%2C%5CnIR7%3DVR7/R7%2C%5CnIR8%3DVR8/R8%2C%5CnIR9%3DVR9/R9%2C%5CnIR10%3DVR10/R10%2C%5CnIR11%3DVR11/R11%2C%5CnIR12%3DVR12/R12%2C%5CnIR13%3DVR13/R13%2C%5CnIR14%3DVR14/R14%2C%5CnIR15%3DVR15/R15%22%2C%20%22Constraints%22%3A%20%22Ic1%3DIR1%2C%5CnIc2%3DIR2%2C%5CnIc3%3DIR3%2C%5CnIc4%3DIR4%2C%5CnIc5%3DIR5%2C%5CnIc6%3DIR6%2C%5CnIc7%3DIR7%2C%5CnIc8%3DIR8%2C%5CnIc9%3DIR9%2C%5CnIc10%3DIR10%2C%5CnIc11%3DIR11%2C%5CnIc12%3DIR12%2C%5CnIc13%3DIR13%2C%5CnIc14%3DIR14%2C%5CnIc15%3DIR15%2C%5CnVR1%3DVs1-Vc1%2C%5CnVR2%3DVs2-Vc2%2C%5CnVR3%3DVs3-Vc3%2C%5CnVR4%3DVs4-Vc4%2C%5CnVR5%3DVs5-Vc5%2C%5CnVR6%3DVs6-Vc6%2C%5CnVR7%3DVs7-Vc7%2C%5CnVR8%3DVs8-Vc8%2C%5CnVR9%3DVs9-Vc9%2C%5CnVR10%3DVs10-Vc10%2C%5CnVR11%3DVs11-Vc11%2C%5CnVR12%3DVs12-Vc12%2C%5CnVR13%3DVs13-Vc13%2C%5CnVR14%3DVs14-Vc14%2C%5CnVR15%3DVs15-Vc15%22%2C%20%22OutputVars%22%3A%20%22Vc1%2C%20Vc2%2C%20Vc3%2C%20Vc4%2C%20Vc5%2C%20Vc6%2C%20Vc7%2C%20Vc8%2C%20Vc9%2C%20Vc10%2C%20Vc11%2C%20Vc12%2C%20Vc13%2C%20Vc14%2C%20Vc15%22%7D). This shows how the page can be scrolled horizontally to view the entire equation.

> 5. Where do you document the "Upload a photo" functionality of the web interface? It's unclear what this does or what the photo should be.

I noticed a bug related to this, which is now fixed. This image is displayed for reference while entering equations on the "Enter System Equations" pane.

> 6. For each example, consider a brief description of what system is being modeled.

This has been added.

## WEB interface tutorial

> 1. The tutorial would benefit from making clear which parts of the tutorial describe what a user must do to prepare input versus describing what the StateMint software does. It seems that the user must manually derive elemental, continuity, and compatibility equations, and the software uses these equations to provide state and output equations. This could be made clearer.

The tutorial has been updated.

> 2. There are numbers on the figures in the Normal Tree section (e.g., 1, 2, 3, 4), but you do not describe the meaning of these numbers.

This has been added.

> 3. The writing would benefit from the use of the active voice instead of the passive voice.

I will be working on this over the next couple of days.

# Python

> It's atypical to put all the code in \_\_init\_\_.py. Please create more descriptive Python files.

The Python library has been moved to a more descriptive filename.

## Example.ipynb

> 1. The first paragraph seems incorrect (there are 3 methods for using StateMint, not only 2).

You are correct. This has been changed.

> 2. Out[5] shows $(t)$ on its own. Placing $t$ in parentheses like this seems odd; is there an error?

That is really strange. Running the Jupyter notebook on my machine this issue is not present. I have contacted GitHub about this issue.

# Mathematica

> Please state the version of Mathematica with which you tested StateMint.

This has been added to the Mathematica folder readme.
