### 2019-02-25 #44 @moorepants
> Would the addition to the paper of an explanation of the algorithm used to find the differential equations be enough to satisfy this requirement?

The pedagogical soundness is asking for you to justify why your software/lesson is good for teaching and learning. Backing this with some educational research evidence is ideal.
