### 2019-02-26 #44 @CameronDevine
@moorepants and @chrisdembia, the following statement is already included in the paper.

> Finally, the set of equations can be reduced to a system of first-order differential equations
in state and input variables and their time-derivatives, alone. It is in these last two steps,
especially the very last, that a student, manually reducing the set of equations, will often make some minor mistake. This is typically of a “book keeping” variety that, if it
teaches the student anything, it is not system dynamics. Instead, the student can be
easily discouraged and confused about where they have made their mistake. Fortunately,
the software tools presented here will automate the algebraic reduction. These will allow
students to focus on understanding the process of dynamic system modeling.

Prof. Rico Picone and Prof. Joseph Garbini both have many years of experience teaching system dynamics using linear graph methods. They both agree that this is the area where most students struggle. We could add more to the paper to explain this.

We also could distribute a survey to Prof. Rico Picone's current students, possibly including some of the questions below.

> 1. At which step in the state-space modeling algorithm do you most frequently make a mistake? [list steps of algorithm]
> 2. If there was an effective tool that automated step X [the thing StateMint does], would you find that it helps your learning process? If yes, why? If no, why not?
> 3. Consider the tool at statemint.stmartin.edu. Try an example. Do you think this is an effective tool for the automation of step X?

Which of these do you think would best show the pedagogical soundness of this tool?
