### 2019-02-26 #44 @moorepants
@chrisdembia @gboeing Can you both comment on @CameronDevine's last post? I'll defer to the reviewers on deciding how to handle the pedagogical soundness requirement. I can help provide facts and clarifications.
