### 2019-02-26 1632f27
added a requirements.txt file. Closes #9
