+++ python/requirements.txt
@@ -1,0 +1,1 @@
+sympy>=0.7.3
