### 2019-02-26 #44 @gboeing
In my opinion, a survey is not *necessary* but it would be nice. Certainly more explanation regarding your comment ("Prof. Rico Picone and Prof. Joseph Garbini both have many years of experience teaching system dynamics using linear graph methods. They both agree that this is the area where most students struggle. We could add more to the paper to explain this.") would be helpful to elucidate too.
