### 2019-03-21 #44 @labarba
Hi everybody!
We don't require formal evaluation of the learning module, but we expect the author to describe their instructional design and the pedagogy behind it.

I see both reviewers have some items of their review lists unchecked. Are you working on addressing some of the reviewer comments, @CameronDevine?
