### 2019-03-22 70aecc3
changed branch to edge to reduce confusion, made a few wording adjustments, fixed a couple typos
