### 2019-03-23 9ef213d
added rico's changes to the website tutorial
