+++ web/HTML/tutorial/tutorial.md
@@ -2,7 +2,7 @@
 
 The purpose of this tutorial is to give the necessary background information on how to use the StateMint software to those who have some experience with system dynamics but may need a refresher, or are not familiar with the linear graph method.
 To learn more about this subject consider reading _System Dynamics: An Introduction_ by Rowell and Wormley.
-Also available are course [notes](http://ricopic.one/dynamic_systems/) for a class at St. Martin's university taught by Prof. Rico Picone, along with [recorded](https://www.youtube.com/watch?v=Fd1C-abrpmg&index=22&list=PLtuwVtW88fOcFdJ9xOBn0T5ta_XPMKHKz) lectures.
+Also available are course [notes](http://ricopic.one/dynamic_systems/) for a class at St. Martin's University taught by Prof. Rico Picone, along with [recorded](https://www.youtube.com/watch?v=Fd1C-abrpmg&index=22&list=PLtuwVtW88fOcFdJ9xOBn0T5ta_XPMKHKz) lectures.
 
 ## Background Information
 
@@ -27,12 +27,12 @@ A list of common system types and their through-variables and across-variables c
 
 ### Elemental Equations
 
-Within each energy domain, there are multiple element types which can be used to model a system in that domain.
-Across energy domains, these elements can be grouped into three distinct types.
+Within each energy domain, there are multiple element-types which can be used to model a system in that domain.
+Across energy domains, these elements can be grouped into three distinct-types.
 
-#### A Type Elements
+#### A-Type Elements
 
-A type elements are energy storage elements which relate the rate of change of the across-variable to the through-variable of the element.
+A-type elements are energy storage elements which relate the rate of change of the across-variable to the through-variable of the element.
 
 | System Type   | Element             | Elemental Equation         | Parameter
 | ------------- | ------------------- | -------------------------- | ---------
@@ -42,9 +42,9 @@ A type elements are energy storage elements which relate the rate of change of t
 | Fluid         | Fluid Capacitor     | $Q=C_f\frac{dP}{dt}$       | Fluid Capacitance, $C_f$
 | Thermal       | Thermal Capacitance | $q=C_t\frac{dT}{dt}$       | Thermal Capacitance, $C_t$
 
-#### T Type Elements
+#### T-Type Elements
 
-T type elements are energy storage elements which relate the rate of change of the through-variable to the across-variable of the element.
+T-type elements are energy storage elements which relate the rate of change of the through-variable to the across-variable of the element.
 
 | System Type   | Element          | Elemental Equation                     | Parameter
 | ------------- | ---------------- | -------------------------------------- | ---------
@@ -53,9 +53,9 @@ T type elements are energy storage elements which relate the rate of change of t
 | Electrical    | Inductor         | $v=L\frac{di}{dt}$                     | Inductance, $L$
 | Fluid         | Inertance        | $P=I_f\frac{dQ}{dt}$                   | Fluid Inertance, $I_f$
 
-#### D Type Elements
+#### D-Type Elements
 
-D type elements are strictly dissipative with a linear relationship between through-variables and across-variables.
+D-type elements are strictly dissipative with a linear relationship between through-variables and across-variables.
 
 | System Type   | Element            | Elemental Equation         | Parameter
 | ------------- | ------------------ | -------------------------- | ---------
@@ -101,37 +101,37 @@ The following elements can be modeled as gyrators,
 
 #### Sources
 
-The final type of element are the ideal sources.
+The final type of element is the _ideal source_.
 Ideal sources provide (potentially arbitrary) external power to a system by specifying either its across-variable or through-variable.
 The former are called across-variable sources and the latter through-variable sources.
 
 ### Linear Graphs
 
-A linear graph is a diagram of the system using "nodes" connected by "branches"
+A linear graph is a diagram of a system that represents its topology via nodes and edges (lines).
[…]
