### 2019-03-23 e12e517
added more italics keywords to the tutorial
