+++ tutorial.md
@@ -11,9 +11,9 @@ This method of finding a differential equation of a dynamic system works with mu
 ### System Types
 
 In each energy domain, a through-variable, ![$v$](http://latex.codecogs.com/svg.latex?v), and an across-variable, ![$f$](http://latex.codecogs.com/svg.latex?f), is defined.
-A through-variable is a variable that generally has the same value at each element terminal.
+A _through-variable_ is a variable that generally has the same value at each element terminal.
 It corresponds to a physical quantity that would be measured flowing through an element.
-An across-variable is a variable that generally has a different value at each element terminal.
+An _across-variable_ is a variable that generally has a different value at each element terminal.
 It corresponds to a physical quantity that would be measured across an element or relative to some reference.
 A list of common system types and their through-variables and across-variables can be found below.
 
@@ -32,7 +32,7 @@ Across energy domains, these elements can be grouped into three distinct-types.
 
 #### A-Type Elements
 
-A-type elements are energy storage elements which relate the rate of change of the across-variable to the through-variable of the element.
+_A-type_ elements are energy storage elements which relate the rate of change of the across-variable to the through-variable of the element.
 
 | System Type   | Element             | Elemental Equation         | Parameter
 | ------------- | ------------------- | -------------------------- | ---------
@@ -44,7 +44,7 @@ A-type elements are energy storage elements which relate the rate of change of t
 
 #### T-Type Elements
 
-T-type elements are energy storage elements which relate the rate of change of the through-variable to the across-variable of the element.
+_T-type_ elements are energy storage elements which relate the rate of change of the through-variable to the across-variable of the element.
 
 | System Type   | Element          | Elemental Equation                     | Parameter
 | ------------- | ---------------- | -------------------------------------- | ---------
@@ -55,7 +55,7 @@ T-type elements are energy storage elements which relate the rate of change of t
 
 #### D-Type Elements
 
-D-type elements are strictly dissipative with a linear relationship between through-variables and across-variables.
+_D-type_ elements are strictly dissipative with a linear relationship between through-variables and across-variables.
 
 | System Type   | Element            | Elemental Equation         | Parameter
 | ------------- | ------------------ | -------------------------- | ---------
@@ -68,7 +68,7 @@ D-type elements are strictly dissipative with a linear relationship between thro
 #### Transformers
 
 In dynamic systems, energy can flow between multiple domains.
-Transformers are one model for relating the through-variable and across-variable of two energy domains to each other,
+_Transformers_ are one model for relating the through-variable and across-variable of two energy domains to each other,
 
 ![$\left[\begin{array}{c}v_1\\f_1\end{array}\right]=\left[\begin{array}{cc}TF&0\\0&-1/TF\end{array}\right]\left[\begin{array}{c}v_2\\f_2\end{array}\right]$](http://latex.codecogs.com/svg.latex?%5Cleft%5B%5Cbegin%7Barray%7D%7Bc%7Dv_1%5C%5Cf_1%5Cend%7Barray%7D%5Cright%5D%3D%5Cleft%5B%5Cbegin%7Barray%7D%7Bcc%7DTF%260%5C%5C0%26-1/TF%5Cend%7Barray%7D%5Cright%5D%5Cleft%5B%5Cbegin%7Barray%7D%7Bc%7Dv_2%5C%5Cf_2%5Cend%7Barray%7D%5Cright%5D)
 
@@ -87,7 +87,7 @@ A few of these are listed below,
 
 #### Gyrators
 
-Gyrators are another tool used to model energy flow between domains.
[…]
