### 2019-03-23 4ab606d
there were a few things that didn't need to be installed for the deploy stage of the ci
