+++ .travis.yml
@@ -38,11 +38,7 @@ jobs:
       if: branch = master
       python: 2.7
       install:
-        - pip install piexif selenium awscli;
-          cd web;
-          wget https://chromedriver.storage.googleapis.com/2.44/chromedriver_linux64.zip;
-          unzip chromedriver_linux64.zip;
-          PATH=$PATH:$(pwd);
+        - pip install awscli;
           aws configure set region $AWS_DEFAULT_REGION;
       script:
         - make trackingID=$TRACKING_ID;
