### 2019-03-23 6c040fb
telling the web makefile to quietly zip the necessary python files
