### 2019-03-23 522a648
fixing the ci config file
