### 2019-03-23 CameronDevine/StateModelRnD#10 @CameronDevine
I will use Zenodo to get a DOI number once all the other issues have been solved and the paper is ready to be published.
