### 2019-03-25 #44 @CameronDevine
@moorepants I believe I have fixed all but two of the issues brought up by @chrisdembia and @gboeing.

1. I will be adding more to the statement of need in the paper later today, and I should have a new version which can be reviewed in the next couple of days.

2. camerondevine/StateMint#10 is still open, and I will use Zenodo to get a DOI number for the source code after the other changes have been approved by @chrisdembia and @gboeing.
