### 2019-03-30 4d6ef70
added more time before the timeout for the examples to run
