### 2019-04-03 3db6041
minor edits to early wording
