+++ paper/paper.md
@@ -26,32 +26,33 @@ bibliography: paper.bib
 # Abstract
 
 StateMint is a set of software tools that reduce sets of dynamic equations and their constraints to a state-space model and related dynamic system model formulations.
-These tools are especially useful for the student of system dynamics, many of whom can become lost in this algebraic reduction.
+These tools are especially useful for students of system dynamics, many of whom can become lost in this algebraic reduction.
 StateMint includes a Mathematica package, a Python package, and a web interface that is built as a layer on top of the Python package.
 
 # Introduction
 
-When deriving a system's state-space model&mdash;that is, the vector state (differential) equation and the vector output (algebraic) equation&mdash;one begins by forming scalar equations for each lumped-parameter element, describing its dynamics.
-Lumped-parameter elements, or more generally lumped-parameter modeling, uses the assumption that power enters and exits each element of the system in a finite number of ports, usually one.
-The next step is to form a set of $N$ constraint equations that describe the topology of the system defined by the interconnection of the $N$ elements.
+When deriving a system's state-space model&mdash;that is, the vector state (differential) equation and the vector output (algebraic) equation&mdash;in the manner of Rowell and Wormley ([@rowell1997]; also see the [tutorial](https://github.com/CameronDevine/StateMint/blob/master/tutorial.md)), one begins by forming scalar equations for each lumped-parameter element, describing its dynamics; we call these "elemental equations."
+We assume power enters and exits each element through a finite number of ports, usually one.
+Let the total number of ports in a given system be $N$, which is equal to the number of elemental equations.
+$N$ constraint equations that describe the topology of the system can be defined by the interconnection of the elements.
 A set of $2N$ differential and algebraic equations and $2N$ unknown variables result.
-If properly constructed (e.g. with the linear graph technique [@rowell1997] summarized in the [tutorial](https://github.com/CameronDevine/StateMint/blob/master/tutorial.md)), $N$ of the unknown variables can be immediately eliminated through direct substitution.
+If properly constructed [@rowell1997], $N$ of the unknown variables can be immediately eliminated through direct substitution.
 Finally, the set of equations can be reduced to a system of first-order differential equations in state and input variables and their time-derivatives, alone.
 
 ## Statement of Need
 
-In our experience,^[Collectively we have almost 30 years of experience teaching system dynamics.] a student manually reducing the set of equations, will often make some minor mistake during the tedious process of the last two steps above.
+In our experience,^[Collectively, we have almost 30 years of experience teaching system dynamics.] a student manually reducing the set of equations, will often make some minor mistake during the tedious process of the last two steps above.
 This is typically of a "book keeping" variety that, if it teaches the student anything, it is not system dynamics.
 Instead, the student can be easily discouraged and confused about where they have made their mistake.
 Furthermore, the tedious nature of the reduction can dampen a student's desire to explore interesting problems.
-The software tools presented here will automate this algebraic reduction.
-They will allow students to focus on understanding the process of dynamic system modeling and to experience a wider breadth of problems.
+The software tools presented here automate this algebraic reduction.
+They allow students to focus on understanding the process of dynamic system modeling and to experience a wider breadth of problems.
 
 ## Software Tools
 
-Utilizing the advanced symbolic mathematics capabilities of Mathematica, a package was written to determine the dynamic system model.
-However, this requires students to install and learn (at least) the basics of Mathematica.
-To mitigate this problem, a web interface was designed to allow students to use this tool without any knowledge of programming by allowing equations to be input with notation similar to that of many scientific calculators.
+Originally, utilizing the advanced symbolic mathematics capabilities of Mathematica, a package was written to perform these algebraic reductions.
+However, it requires students to install and learn the basics of Mathematica.
[…]
