### 2019-04-03 8bd16f7
moved a reference to the tutorial into a footnote
