### 2019-04-03 e6570df
switched from using sympify to parse expression with a custom global dict
