### 2019-04-03 #44 @CameronDevine
@chrisdembia and @gboeing, I believe I have addressed all of your comments. If there is anything outstanding please let me know.
