### 2019-04-03 4dce151
modified the download/save buttons to always be shown
