### 2019-04-03 dcfe2b1
added a unique stack name for the web test to allow parallel testing of different commits
