+++ .travis.yml
@@ -32,8 +32,8 @@ jobs:
           unzip chromedriver_linux64.zip;
           PATH=$PATH:$(pwd);
           aws configure set region $AWS_DEFAULT_REGION;
-      script: make all test stack=StateMint-test;
-      after_script: make remove stack=StateMint-test
+      script: make all test stack=StateMint-test-$TRAVIS_JOB_ID;
+      after_script: make remove stack=StateMint-test-$TRAVIS_JOB_ID
     - stage: deploy
       if: branch = master
       python: 2.7
