### 2019-04-03 19771b4
modified the makefile so that a bucket is made for every stack name
