### 2019-04-03 eb7e500
changed the ci config to also remove the aws bucket after it runs
