### 2019-04-03 2aeb6a6
converted the stack name to lowercase for the bucket name
