### 2019-04-05 #44 @labarba
Hi @gboeing — if you're ready to recommend publication, can you check off your remaining items in the checklist? Cheers!
