### 2019-04-07 #44 @chrisdembia
The checklist above contains an item " A statement of need: Do the authors clearly state the need for this software and who the target audience is?" under **Documentation**. I found a statement of need in the paper but not in the documentation. Where can I find such a statement?
