### 2019-04-07 #44 @labarba
The request for a Statement of Need refers to the JOSE paper.
