### 2019-04-07 #44 @labarba
I think this could be an editing mistake on creating the checklists. What would be a reason to request this in the software documentation? I rather think it's enough to have it in the paper.
