### 2019-04-08 #44 @moorepants
@CameronDevine I recommend your article for publishing. Can you please mint a DOI? (https://github.com/CameronDevine/StateMint/issues/10)
