### 2019-04-08 #44 @CameronDevine
@moorepants would you mind if I rolled the version number to 2.1.0 first?
