### 2019-04-08 #44 @moorepants
Yes, you can make the version number anything you want. Please update in the paper once changed.
