### 2019-04-08 #44 @labarba
@CameronDevine : You can try running `@whedon set v2.1.0 as version` here, but I'm not sure if authors have permission to run that command.That will update the version in this issue and the paper's side bar.
