+++ web/aws/api-template.yml
@@ -1,7 +1,7 @@
 ---
 swagger: "2.0"
 info:
-  version: "2.0.0"
+  version: "2.1.0"
   title: "StateMint"
 basePath: "/StateMint"
 schemes:
