### 2019-04-08 #44 @CameronDevine
@moorepants can you change the version to 2.1.0? I also uplaoded the code to Zendodo, (doi: 10.5281/zenodo.2633330).
