### 2019-04-09 CameronDevine/StateModelRnD#10 @moorepants
DOI is there now: 10.5281/zenodo.2633330
