### 2019-04-09 #44 @labarba
Congratulations, @CameronDevine, your JOSE paper is now published!

Big thanks to the editor: @moorepants, and the reviewers: @chrisdembia, @gboeing — your contribution is so important 🙏
