### 2019-04-10 #44 @CameronDevine
@moorepants, You asked about the benefits of Linear Graphs early in the review process. I forgot about your question until I was reviewing some notes today.

The main benefit of Linear Graphs is the resulting diagram is almost identical to an electrical circuit diagram. Since students often learn to read electrical circuit diagrams in physics, we can leverage this existing knowledge.

If you have any other questions, feel free to email me, camdev@uw.edu.
